Split `GraphNodeType` into its own file.
